Avoid throwing exceptions from stol, stold, stoi etc. functions.
Summary:
We used the std:stol, std:stoll and std::stold functions throughout our codebase and in
many cases didn't handle the exceptions appropriately. We already had wrapper functions like
CheckedStol etc. that handled errors properly and returned an appropriate status message.

I've replaced most of these instances across the codebase with the checked versions.

Test Plan:
1) Added several unit tests that would crash the tserver before this change and pass
after this change.
2) jenkins.

Reviewers: neil, robert, mihnea, hector

Reviewed By: hector

Subscribers: ybase, yql

Differential Revision: https://phabricator.dev.yugabyte.com/D3148

+++ src/yb/ql/ptree/pt_property.h
@@ -76,7 +76,7 @@ class PTProperty : public TreeNode {
 
   static CHECKED_STATUS GetDoubleValueFromExpr(PTExpr::SharedPtr expr,
                                                const string& property_name,
-                                               double *val);
+                                               long double *val);
 
   static CHECKED_STATUS GetBoolValueFromExpr(PTExpr::SharedPtr expr,
                                              const string& property_name,
