+++ src/yb/util/decimal.h
@@ -101,7 +101,7 @@ class Decimal {
   std::string ToString() const;
   // Note: We are using decimal -> string -> double using std::stod() function.
   // In future, it may be better to write a direct conversion function.
-  CHECKED_STATUS ToDouble(double* int64_value) const;
+  CHECKED_STATUS ToDouble(long double* int64_value) const;
 
   // Note: The length of the varint is limited by kDefaultMaxLength by default to make sure we don't
   // get stuck with large exponents. May be overriden.
